+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/ComisionesRepository.java
@@ -41,7 +41,7 @@ public interface ComisionesRepository extends JpaRepository<MovimientoComision, 
 	@Modifying
 	@Query("DELETE FROM MovimientoComision mc WHERE mc.id IN (SELECT mcon.id FROM MovimientoConciliacion mcon WHERE mcon.id in :idsComisiones AND mcon.nuevo = true AND mcon.idConciliacion = :idConciliacion)")
 	public void deleteByIdsAndIdConciliacion(@Param("idsComisiones") final List<Integer> idsComisiones,
-			@Param("idConciliacion") final Integer idConciliacion);
+			@Param("idConciliacion") final Long idConciliacion);
 
 	/**
 	 * Verfica si es posible eliminar una comision
@@ -52,7 +52,7 @@ public interface ComisionesRepository extends JpaRepository<MovimientoComision, 
 	 */
 	@Query("SELECT mc.id FROM MovimientoComision mc WHERE mc.id IN :idsComisiones AND mc.id IN (SELECT mcon.id FROM MovimientoConciliacion mcon WHERE mcon.id in :idsComisiones AND mcon.nuevo = true AND mcon.idConciliacion = :idConciliacion)")
 	public List<Long> verifyById(@Param("idsComisiones") final List<Integer> idsComisiones,
-			@Param("idConciliacion") final Integer idConciliacion);
+			@Param("idConciliacion") final Long idConciliacion);
 
 	/**
 	 * Regresa el total de movimientos de pagos que estan entre las fechas
@@ -64,7 +64,7 @@ public interface ComisionesRepository extends JpaRepository<MovimientoComision, 
 	 */
 	@Query("SELECT COUNT(mp.id) FROM MovimientoPago mp INNER JOIN MovimientoConciliacion mc ON mp.id = mc.id WHERE mc.createdDate BETWEEN :fechaDesde AND :fechaHasta AND mc.idConciliacion = :idConciliacion")
 	public Long findTransaccionesPagosByFechasAndIdConciliacion(@Param("fechaDesde") final Date fechaDesde,
-			@Param("fechaHasta") final Date fechaHasta, @Param("idConciliacion") Integer idConciliacion);
+			@Param("fechaHasta") final Date fechaHasta, @Param("idConciliacion") Long idConciliacion);
 
 	/**
 	 * Regresa el id de un movimiento comision su id de conciliacion asociado y el
@@ -76,7 +76,7 @@ public interface ComisionesRepository extends JpaRepository<MovimientoComision, 
 	 */
 	@Query("SELECT COUNT(md.id) AS countId FROM MovimientoConciliacion mc INNER JOIN MovimientoDevolucion md ON md.id = mc.id WHERE mc.createdDate BETWEEN :fechaDesde AND :fechaHasta AND mc.idConciliacion = :idConciliacion")
 	public Map<String, Object> findDataByFechasAndIdConciliacion(@Param("fechaDesde") final Date fechaDesde,
-			@Param("fechaHasta") final Date fechaHasta, @Param("idConciliacion") Integer idConciliacion);
+			@Param("fechaHasta") final Date fechaHasta, @Param("idConciliacion") Long idConciliacion);
 
 	/**
 	 * Regresa la suma de los movimientos tipo COMISION y los de tipo IVA_COMISION
@@ -96,7 +96,7 @@ public interface ComisionesRepository extends JpaRepository<MovimientoComision, 
 	 * @return
 	 */
 	@Query("SELECT new mx.com.nmp.pagos.mimonte.dto.conciliacion.ComisionDTO(mc.id, mc.fechaOperacion, mc.fechaCargo, mc.monto, mc.descripcion, mcon.nuevo) FROM MovimientoComision mc INNER JOIN MovimientoConciliacion mcon ON mc.id = mcon.id AND mcon.idConciliacion = :folio")
-	public List<ComisionDTO> findByFolio(@Param("folio") final Integer folio);
+	public List<ComisionDTO> findByFolio(@Param("folio") final Long folio);
 
 	/**
 	 * Regresa un movimiento comision por id
@@ -136,7 +136,7 @@ public interface ComisionesRepository extends JpaRepository<MovimientoComision, 
 	 * @return
 	 */
 	@Query(nativeQuery = true, value = "SELECT CASE WHEN (SELECT COUNT(mc.id) FROM to_movimiento_comision mc INNER JOIN to_movimiento_conciliacion mcon ON mc.id = mcon.id WHERE mc.id IN :ids AND mcon.id_conciliacion = :folio) = :tam THEN 1 ELSE 0 END")
-	public Object checkIfFolioAndIdsRelationshipExit(@Param("folio") final Integer folio,
+	public Object checkIfFolioAndIdsRelationshipExit(@Param("folio") final Long folio,
 			@Param("ids") final List<Integer> ids, @Param("tam") final Integer tam);
 
 }
