+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/MovimientoComisionRepository.java
@@ -41,7 +41,7 @@ public interface MovimientoComisionRepository extends JpaRepository<MovimientoCo
 	 * @return
 	 */
 	@Query("SELECT mco FROM MovimientoComision mco INNER JOIN MovimientoConciliacion mc ON mc.id = mco.id WHERE mco.idConciliacion = :folio")
-	public List<MovimientoComision> findByIdConciliacion(@Param("folio") Integer folio);
+	public List<MovimientoComision> findByIdConciliacion(@Param("folio") Long folio);
 
 	/**
 	 * Regresa un movimiento comision por el id de movimiento estado cuenta
@@ -72,7 +72,7 @@ public interface MovimientoComisionRepository extends JpaRepository<MovimientoCo
 	 * @return
 	 */
 	@Query(nativeQuery = true, value = "SELECT CASE WHEN ((SELECT COUNT(mc.id) FROM to_movimiento_comision mc INNER JOIN to_movimiento_conciliacion mcon ON mc.id = mcon.id WHERE mc.id IN (68,69) AND mcon.id_conciliacion = 1) = (SELECT 2)) THEN 1 ELSE 0 END")
-	public Object checkIdsAndFolioRelationship(@Param("folio") final Integer folio,
+	public Object checkIdsAndFolioRelationship(@Param("folio") final Long folio,
 			@Param("ids") final List<Integer> ids, @Param("tam") final Integer tam);
 
 }
