+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/MovimientoEstadoCuentaRepository.java
@@ -32,7 +32,7 @@ public interface MovimientoEstadoCuentaRepository extends PagingAndSortingReposi
 	 * @return
 	 */
 	@Query("SELECT COUNT(mm.id) FROM MovimientoEstadoCuenta mm INNER JOIN EstadoCuenta ec ON mm.idEstadoCuenta = ec.id INNER JOIN Reporte r ON r.id = ec.idReporte WHERE r.conciliacion.id = :folioConciliacion")
-	public Long countMovimientos(@Param("folioConciliacion") final Integer folioConciliacion);
+	public Long countMovimientos(@Param("folioConciliacion") final Long folioConciliacion);
 
 	/**
 	 * Regresa una lista de objetos de tipo MovimientoEstadoCuentaDBDTO relacionados
@@ -44,7 +44,7 @@ public interface MovimientoEstadoCuentaRepository extends PagingAndSortingReposi
 	 * @return
 	 */
 	@Query("SELECT new mx.com.nmp.pagos.mimonte.dto.conciliacion.MovimientoEstadoCuentaDBDTO(mm.id, mm.fechaOperacion, mm.concepto, mm.tipoMovimiento, mm.importe, ec.cabecera.saldoInicial, ec.totalesAdicional.saldoFinal) FROM MovimientoEstadoCuenta mm INNER JOIN EstadoCuenta ec ON mm.idEstadoCuenta = ec.id INNER JOIN Reporte r ON r.id = ec.idReporte WHERE r.conciliacion.id = :folioConciliacion")
-	public List<MovimientoEstadoCuentaDBDTO> listMovimientos(@Param("folioConciliacion") final Integer folioConciliacion
+	public List<MovimientoEstadoCuentaDBDTO> listMovimientos(@Param("folioConciliacion") final Long folioConciliacion
 			/*,Pageable pageable*/);
 
 	/**
@@ -65,6 +65,6 @@ public interface MovimientoEstadoCuentaRepository extends PagingAndSortingReposi
 	public List<MovimientoEstadoCuenta> findByReporteAndClaveLeyendaIn(@Param("reporteId") final Integer reporteId, @Param("clavesLeyenda") List<String> clavesLeyenda);
 
 	@Query("SELECT mm FROM MovimientoEstadoCuenta mm INNER JOIN EstadoCuenta ec ON mm.idEstadoCuenta = ec.id INNER JOIN Reporte r ON r.id = ec.idReporte INNER JOIN r.conciliacion con WHERE con.id = :idConciliacion")
-	public List<MovimientoEstadoCuenta> findByConciliacion(@Param("idConciliacion") final Integer idConciliacion);
+	public List<MovimientoEstadoCuenta> findByConciliacion(@Param("idConciliacion") final Long idConciliacion);
 
 }
